TST: Move log_queue to conftest
